+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add the `--has-license {}` filter to `ct az devops license entitlement user update` cli
 - Add specifying multiple `--user {}` to `ct az devops license entitlement user update` cli
 - Clean up CLI types to use `AzureDevOpsUserArgument` instead of previous matcher duplicate type
+- Add `--log-file-filter` to independently control structured log-file verbosity
 
 # v0.36.0